Move hardware initialization out of europi.py and into europi_hardware.py. Allow aborting wifi connection attempts during startup by pressing either button. Add a note to the wifi startup screen prompting B1 to cancel
